Moved dungeon generator into its own folder. Changed a few imports.
