+++ load.py
@@ -9,8 +9,8 @@ from peewee import SqliteDatabase
 from settings import DATABASE_NAME
 from area.config import LEVELS
 from area.models import Dungeon, DungeonObject, DungeonLevel
-from item.models import Item
-from item.components import Inventory
+from items.models import Item
+from items.components import Inventory
 from character.models import Character
 from character.components import CharacterClass
 from character.config import PLAYER
