+++ tests/test_dungeon.py
@@ -5,15 +5,15 @@ python2.7 -m unittest discover tests
 import json
 import unittest
 
+from area.models import Dungeon, DungeonLevel, DungeonObject
+from character.components import CharacterClass
+from character.models import Character
+from items.models import Item
 from peewee import SqliteDatabase
 
 import load
-from area.models import Dungeon, DungeonLevel, DungeonObject
-from item.models import Item
-from character.models import Character
-from character.components import CharacterClass
 from character.config import ORC, TROLL
-from area.dungeon_generator import DungeonGenerator
+from generation.dungeon_generator import DungeonGenerator
 
 database = SqliteDatabase('roguelike_test.db')
 TEST_TABLES = [Dungeon, DungeonObject, DungeonLevel, Item, Character, CharacterClass]
